Pare site down to mostly text
Currently, the site doesn't even look very good. It did at one point but...that's somehow changed. Anyway, this gets rid of the whole 3-column idea, my avatar, links to twitter and my irc handle and puts things into a bare-minimum state. My concern with a website/portfolio type thing like this is that it can get out-of-date really fast. Hopefully what is left with this change won't go out of date, at least any time soon. And i kinda like the minimalism as a style, too. I may do more with the styles later to spice it up a bit, but it's good for now.